% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit_branch.R
\name{fgit_branch}
\alias{fgit_branch}
\title{List all branch from a repository}
\usage{
fgit_branch(repo)
}
\arguments{
\item{repo}{The repository's name from GitHub.com}
}
\value{
A vector of branches
}
\description{
List all branch from a repository
}
\examples{
\dontrun{
fgit_branch("https://github.com/FastGitORG/fgitR")
fgit_branch("FastGitORG/fgitR")
}
}
