% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit.R
\name{fgit_clone}
\alias{fgit_clone}
\title{Git Clone by FastGit}
\usage{
fgit_clone(repo, overwite = FALSE, verbose = TRUE)
}
\arguments{
\item{repo}{The repo name from GitHub.com}

\item{overwite}{Overwrite the exist directories. Default is `FALSE`}

\item{verbose}{Verbose logs. Default is `TRUE`}
}
\description{
Git Clone by FastGit
}
\examples{

fgit_clone("womeimingzi11/fgitR", overwrite = TRUE)

}
