% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit_clone.R
\name{fgit_clone_from_url}
\alias{fgit_clone_from_url}
\title{Git Clone by FastGit's URL}
\usage{
fgit_clone_from_url(
  fgit_url,
  target_dir,
  branch = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{target_dir}{The output directory}

\item{branch}{Clone from which branch.}

\item{overwrite}{Overwrite the exist directories. Default is `FALSE`}

\item{verbose}{Verbose logs. Default is `TRUE`}

\item{repo}{The repository's name from GitHub.com}
}
\description{
Git Clone by FastGit's URL
}
