% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit.R
\name{fgit_clone_repo}
\alias{fgit_clone_repo}
\title{Git Clone from Repo by FastGit}
\usage{
fgit_clone_repo(repo, dir = tempdir(), overwite = FALSE, verbose = TRUE)
}
\arguments{
\item{repo}{The repo name from GitHub.com}

\item{overwite}{Overwrite the exist directories. Default is `FALSE`}

\item{verbose}{Verbose logs. Default is `TRUE`}
}
\description{
Git Clone from Repo by FastGit
}
\examples{

fgit_clone_repo("womeimingzi11/fgitR", overwrite = TRUE)

}
