% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit.R
\name{fgit_clone_url}
\alias{fgit_clone_url}
\title{Git Clone from URL by FastGit}
\usage{
fgit_clone_url(url, dir = tempdir(), overwite = FALSE, verbose = TRUE)
}
\arguments{
\item{url}{The repo name from GitHub.com}

\item{overwite}{Overwrite the exist directories. Default is `FALSE`}

\item{verbose}{Verbose logs. Default is `TRUE`}
}
\description{
Git Clone from URL by FastGit
}
\examples{

fgit_clone_url("https://hub.fastgit.org/womeimingzi11/amapGeocode.git", overwrite = TRUE)

}
