% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgitR_install.R
\name{fgitR_install}
\alias{fgitR_install}
\title{Install package from GitHub by FastGit accelerating}
\usage{
fgitR_install(repo)
}
\arguments{
\item{url}{The repository name or url from GitHub.com}
}
\description{
Install package from GitHub by FastGit accelerating
}
\examples{
fgitR_install("https://github.com/womeimingzi11/fgitR")
fgitR_install("womeimingzi11/fgitR")
}
