% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{git_repo_extract}
\alias{git_repo_extract}
\title{Extract the Repository Name from URL or Name}
\usage{
git_repo_extract(url_or_repo)
}
\arguments{
\item{url_or_repo}{URL or Name from GitHub.com}
}
\value{
Character like user/repository
}
\description{
Extract the Repository Name from URL or Name
}
