% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgitR_clone.R
\name{fgit_clone}
\alias{fgit_clone}
\title{Git Clone from URL by FastGit}
\usage{
fgit_clone(repo, dir = tempdir(), overwite = FALSE, verbose = TRUE)
}
\arguments{
\item{overwite}{Overwrite the exist directories. Default is `FALSE`}

\item{verbose}{Verbose logs. Default is `TRUE`}

\item{url}{The repo name or url from GitHub.com}
}
\description{
Git Clone from URL by FastGit
}
\examples{

fgit_clone("https://github.com.cnpmjs.org/womeimingzi11/womeimingzi11", overwrite = TRUE)
fgit_clone("womeimingzi11/womeimingzi11", overwrite = TRUE)

}
