% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgitR_clone.R
\name{fgit_clone}
\alias{fgit_clone}
\title{Git Clone from URL by FastGit}
\usage{
fgit_clone(
  repo,
  dir = tempdir(),
  overwite = FALSE,
  verbose = TRUE,
  returen_dir = FALSE
)
}
\arguments{
\item{overwite}{Overwrite the exist directories. Default is `FALSE`}

\item{verbose}{Verbose logs. Default is `TRUE`}

\item{returen_dir}{Returen dir as `target_dir` to parent environment?
Somehow useful to catch the temp directory.}

\item{url}{The repository name or url from GitHub.com}
}
\description{
Git Clone from URL by FastGit
}
\examples{

fgit_clone("https://github.com/womeimingzi11/fgitR", overwrite = TRUE)
fgit_clone("womeimingzi11/fgitR", overwrite = TRUE)
}
