% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit_clone.R
\name{fgit_clone}
\alias{fgit_clone}
\title{Git Clone from URL by FastGit}
\usage{
fgit_clone(
  repo,
  dir = tempdir(),
  branch = NULL,
  repo_check = TRUE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{repo}{The repository name or url from GitHub.com}

\item{dir}{The output directory.
`tempdir()` will be used by default.}

\item{branch}{Clone from which branch.}

\item{repo_check}{Check the existence of repository}

\item{overwrite}{Overwrite the exist directories. Default is `FALSE`}

\item{verbose}{Verbose logs. Default is `TRUE`}
}
\value{
No return value, called for side effects
}
\description{
Git Clone from URL by FastGit
}
\examples{
\dontrun{
fgit_clone("https://github.com/womeimingzi11/fgitR", overwrite = TRUE)
fgit_clone("womeimingzi11/fgitR", overwrite = TRUE)
}
}
