% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit_install.R
\name{fgit_install}
\alias{fgit_install}
\title{Install package from GitHub by FastGit accelerating}
\usage{
fgit_install(repo)
}
\arguments{
\item{repo}{The repository name or url from GitHub.com}

\item{overwite}{Overwrite the exist directories. Default is `FALSE`}

\item{branch}{Install from which branch.}
}
\description{
Install package from GitHub by FastGit accelerating
}
\examples{
fgitR_install("https://github.com/womeimingzi11/fgitR")
fgitR_install("womeimingzi11/fgitR")
}
