% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_mk_dir}
\alias{func_mk_dir}
\title{A function to make directory}
\usage{
func_mk_dir(target_dir, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{target_dir}{Path of target Directory}

\item{overwrite}{Overwrite the exist directories. Default is `FALSE`}

\item{verbose}{Verbose logs. Default is `TRUE`}
}
\description{
A function to make directory
}
